Fix for backtranslation of things like http://address.com
This fixes the problem where even though things like
http://address.com or https://address2.com has the no back translate
symbols around it the ‘://’ gets back translated to ‘whstst’  So what
I have done in en-us-g2.ctb is added a rule of:

midword :// 156-34-34

this works for ssh:// http:// ftp:// and all of the things like it. I
have had no problems with this on the braille plus 18. I am not sure
that we should have to do this if there are the no back translation
symbols around the whole address but this fixes the problem.


git-svn-id: http://liblouis.googlecode.com/svn/trunk@1030 24698d54-e344-0410-af24-25bd253c2c7b