Restructured lou_harnessGenerator: for greater readability.
Implementation is identical to the original, except for the
output filename generation. Previously we had a problem if we have two
harness source files that use the same braille table. The second source
would overwrite the harness file output of the first source. Now The
new filename will be the source filename with _source removed.

git-svn-id: http://liblouis.googlecode.com/svn/trunk@1050 24698d54-e344-0410-af24-25bd253c2c7b